disable TestInterval with -race test option
