+++ plugins/inputs/sysstat/sysstat_test.go
@@ -8,7 +8,6 @@ import (
 	"os/exec"
 	"path"
 	"testing"
-	"time"
 
 	"github.com/influxdata/telegraf/testutil"
 )
@@ -303,31 +302,4 @@ dell-xps	5	2016-03-25 16:18:10 UTC	sdb	%util	0.30
 	}
 	// some code here to check arguments perhaps?
 	os.Exit(0)
-}
-
-// TestGatherInterval checks that interval is correctly set.
-func TestGatherInterval(t *testing.T) {
-	// overwriting exec commands with mock commands
-	execCommand = fakeExecCommand
-	defer func() { execCommand = exec.Command }()
-	var acc testutil.Accumulator
-
-	s.interval = 0
-	wantedInterval := 3
-
-	err := s.Gather(&acc)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	time.Sleep(time.Duration(wantedInterval) * time.Second)
-
-	err = s.Gather(&acc)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if s.interval != wantedInterval {
-		t.Errorf("wrong interval: got %d, want %d", s.interval, wantedInterval)
-	}
 }
